requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.3
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
codetiming==1.3.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2
django-celery-beat==2.2.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
isort==5.8.0
kombu==5.0.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
pyrate-limiter==2.3.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
ratelimit==2.2.1
redis==3.5.3
requests==2.25.1
responses==0.13.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
